make DRY create and delete spreadsheet config json file
